# artifacts/open_api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Chat APIs
  version: 0.215.37
  description: API for the Simple Chat, handling chats, conversations, documents, user settings, groups, feedback, prompts, models, safety, and admin functions.
servers:
  - url: /api
    description: API Root
...
paths:
...
  /feedback/submit:
    post:
      tags:
        - Feedback
      summary: Submit feedback for a message
      description: Allows authenticated users to submit feedback (e.g., 'Positive', 'Negative', 'Neutral') on a specific AI message within a conversation. Requires `enable_user_feedback` setting to be true.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messageId
                - conversationId
                - feedbackType
              properties:
                messageId:
                  type: string
                  description: The ID of the AI message receiving feedback.
                conversationId:
                  type: string
                  description: The ID of the conversation containing the message.
                feedbackType:
                  type: string
                  enum: [Positive, Negative, Neutral]
                  description: The type of feedback.
                reason:
                  type: string
                  description: Optional reason or comment for the feedback.
      responses:
        '200':
          description: Feedback submitted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  feedbackId:
                    type: string
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Feature `enable_user_feedback` disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Failed to query messages or save feedback)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /feedback/review:
    get:
      tags:
        - Feedback
        - Admin
      summary: Get feedback items for admin review (paginated)
      description: Retrieves submitted feedback items for administrative review, with pagination and filtering options. Requires admin privileges and `enable_user_feedback` setting.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number.
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            enum: [10, 20, 50]
          description: Items per page.
        - name: type
          in: query
          schema:
            type: string
            enum: [Positive, Negative, Neutral]
          description: Filter by feedback type.
        - name: ack
          in: query
          schema:
            type: string
            enum: ['true', 'false']
          description: Filter by acknowledged status ('true' or 'false').
      responses:
        '200':
          description: List of feedback items retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  feedback:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeedbackItem'
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_count:
                    type: integer
                  total_pages:
                    type: integer
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Not admin or feature disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database query failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /feedback/review/{feedbackId}:
    get:
      tags:
        - Feedback
        - Admin
      summary: Get a single feedback item (Admin)
      description: Fetches the details of a single feedback item by its ID. Needed for editing after pagination. Requires admin privileges and `enable_user_feedback` setting.
      security:
        - bearerAuth: []
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the feedback item to retrieve.
      responses:
        '200':
          description: Feedback item retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackItem'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Not admin or feature disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Feedback item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database read failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Feedback
        - Admin
      summary: Update admin review fields for a feedback item (Admin)
      description: Allows admins to update the review status (`acknowledged`) and associated notes/actions for a specific feedback item. Requires admin privileges and `enable_user_feedback` setting.
      security:
        - bearerAuth: []
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the feedback item to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                acknowledged:
                  type: boolean
                analysisNotes:
                  type: string
                responseToUser:
                  type: string
                actionTaken:
                  type: string
      responses:
        '200':
          description: Feedback item updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Invalid data format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Not admin or feature disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Feedback item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database read or update failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /feedback/retest/{feedbackId}:
    post:
      tags:
        - Feedback
        - Admin
      summary: Retest a prompt from feedback (Admin)
      description: Allows admins to re-run the original user prompt (provided in the request body, typically sourced from the feedback item) against the current AI model configuration to check for improvements. Requires admin privileges and `enable_user_feedback` setting.
      security:
        - bearerAuth: []
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the feedback item context (used for authorization/logging, prompt sent in body).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
              properties:
                prompt:
                  type: string
      responses:
        '200':
          description: Retest completed successfully (returns placeholder response in current code).
          content:
            application/json:
              schema:
                type: object
                properties:
                  retestResponse:
                    type: string
        '400':
          description: Missing prompt in request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Not admin or feature disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Failed to run prompt against AI model)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error that occurred.
...
    FeedbackItem:
      type: object
      description: Represents a single piece of user feedback on an AI message.
      properties:
        id:
          type: string
          format: uuid
          description: Unique ID for the feedback entry.
        userId:
          type: string
          nullable: true
          description: ID of the user who submitted the feedback.
        conversationId: # Added based on implementation
          type: string
          format: uuid
          description: ID of the conversation the feedback relates to.
        messageId: # Added based on implementation
          type: string
          description: ID of the specific AI message the feedback is about.
        prompt:
          type: string
          description: The user prompt that led to the AI response being reviewed.
        aiResponse:
          type: string
          description: The text of the AI response being reviewed.
        feedbackType:
          type: string
          enum: [Positive, Negative, Neutral]
          description: The type of feedback submitted.
        reason:
          type: string
          nullable: true
          description: Optional user comment or reason.
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the feedback was submitted.
        adminReview:
          type: object
          description: Fields related to administrative review of the feedback.
          properties:
            acknowledged:
              type: boolean
              default: false
              description: Has an admin acknowledged/reviewed this feedback?
            analyzedBy:
              type: string
              nullable: true
              description: ID of the admin who reviewed the feedback.
            analysisNotes:
              type: string
              nullable: true
              description: Admin's notes on the feedback analysis.
            responseToUser:
              type: string
              nullable: true
              description: Any response communicated back to the user (if applicable).
            actionTaken:
              type: string
              nullable: true
              description: Description of any corrective action taken.
            reviewTimestamp:
              type: string
              format: date-time
              nullable: true
              description: Timestamp when the admin review fields were last updated.
...
  securitySchemes:
    bearerAuth: # Defines how API clients authenticate
      type: http
      scheme: bearer
      bearerFormat: JWT # Indicates the type of token expected
      description: "Azure AD issued JWT Bearer token required for authenticated endpoints. Include in the request header as 'Authorization: Bearer <YOUR_TOKEN>'."